POC: extract RfidControllers logic in separate app runned at TagBehind
https://github.com/Not-Endurance/not-timing-system/issues/281
