+++ .gitignore
@@ -195,10 +195,6 @@ PublishScripts/
 # NuGet v3's project.json files produces more ignorable files
 *.nuget.props
 *.nuget.targets
-# rfid dlls
-rfid.dll
-rfid_core.dll
-Vup.dll
 
 # Microsoft Azure Build Output
 csx/
