--- src-v2/Judge/NTS.Judge.Blazor/Pages/Setup/RfidTags/RfidTagScanner.razor
@@ -1,54 +1,0 @@
-﻿@using NTS.Judge.Blazor.Ports
-@using NTS.Domain.Setup.Entities
-
-@inherits BindableValueComponent<Tag>
-
-<MudInputLabel Class="mr-4 mt-2 font-size-12">@Localizer.Get("Scan Rfid Tag")</MudInputLabel>
-<MudStack Style="height:80px" Row="true">
-    @if(_awaitingScan)
-    {
-        <MudStack Class="mt-3" >
-            <MudProgressCircular Class="align-self-center" Color="Color.Primary" Size="Size.Small" Indeterminate="true" />
-            <MudText Typo="Typo.overline" Color="Color.Primary">
-                @Localizer.Get("Waiting for Tag Scan")
-            </MudText>
-        </MudStack>
-    }
-    else
-    {
-        if (Value == null)
-        {
-            <MudButton Style="height:32px" Class="mt-3" Color="Color.Primary" Size="Size.Small" Variant="Variant.Filled"
-                       StartIcon="@Icons.Material.Filled.Add" OnClick="AddTag">@Localizer.Get("Add Tag")</MudButton>
-        }
-        else
-        {
-            <MudPaper Style="height:32px" Class="mt-3">
-                <MudText Class="ml-2 mr-2" Typo="Typo.overline"> @Value </MudText>
-            </MudPaper>
-
-            <MudButton Style="height:32px" Class="ml-3 mt-3" Color="Color.Error" Size="Size.Small" Variant="Variant.Filled"
-                       StartIcon="@Icons.Material.Filled.Delete" OnClick="RemoveTag">@Localizer.Get("Remove Tag")</MudButton>
-        }
-    }
-</MudStack>
-
-@code {
-    [Inject]
-    public ITagWrite _tagWriter { get; set; } = default!;
-    [Parameter]
-    public int CombinationNumber { get; set; }
-    public bool _awaitingScan = false;
-
-    public async Task AddTag()
-    {
-        _awaitingScan = true;
-        Value = await _tagWriter.Write(CombinationNumber);
-        _awaitingScan = false;
-    }
-
-    public void RemoveTag()
-    {
-        Value = null;
-    }
-}
