+++ src-v2/Judge/NTS.Judge/Adapters/Behinds/TagWriterBehind.cs
@@ -1,25 +1,41 @@
 ﻿using NTS.Domain.Setup.Entities;
 using NTS.Judge.Blazor.Ports;
-using NTS.Judge.Hardware;
+using System.Diagnostics;
 
-namespace NTS.Judge.Adapters.Behinds;
+namespace NTS.Judge.Adapters;
 
 public class TagWriterBehind : ITagWrite
 {
-    private VupVD67Controller VD67Controller;
-
-    public TagWriterBehind()
+    public async Task<Tag> WriteAsync(int number)
     {
-        VD67Controller = new VupVD67Controller(TimeSpan.FromMilliseconds(100));
-        VD67Controller.Connect();
-    }
+        string output = String.Empty;
+        string error = String.Empty;
+        ProcessStartInfo processStartInfo = new ProcessStartInfo
+        {
+            FileName = @"C:\Work\not-timing-system\src-v2\NTS.TagRfidControllers\bin\Debug\net8.0\NTS.TagRfidControllers.exe",
+            Arguments = number.ToString(),
 
-    public async Task<Tag> Write(int number)
-    {
-        var data = await Task.Run(VD67Controller.Read);
-        var id = data.Substring(9);
-        var tag = new Tag(id, number);
-        VD67Controller.Write(tag.PrepareToWrite());
-        return tag;
+            RedirectStandardOutput = true,
+            RedirectStandardError = true,
+            UseShellExecute = false,
+            CreateNoWindow = true
+        };
+
+        try
+        {
+            using (Process process = Process.Start(processStartInfo))
+            {
+                output = await process.StandardOutput.ReadToEndAsync();
+                error = await process.StandardError.ReadToEndAsync();
+
+                process.WaitForExit();
+            }
+        }
+        catch
+        {
+            // Log error.
+        }
+        string tagId = output.Split(" ").Last();
+        return new Tag(tagId, number);
     }
 }
