+++ src-v2/Judge/NTS.Judge/NTS.Judge.csproj
@@ -18,29 +18,6 @@
     <ProjectReference Include="..\NTS.Judge.Blazor\NTS.Judge.Blazor.csproj" />
   </ItemGroup>
 
-	<ItemGroup>
-		<Content Include="rfid.dll">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</Content>
-		<Content Include="rfid_core.dll">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</Content>
-		<Content Include="Vup.dll">
-			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
-			<ExcludeFromSingleFile>true</ExcludeFromSingleFile>
-		</Content>
-	</ItemGroup>
-
-	<ItemGroup>
-	  <PackageReference Include="Microsoft.Windows.Compatibility" Version="8.0.8" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <Reference Include="Vup">
-	    <HintPath>Vup.dll</HintPath>
-	  </Reference>
-	</ItemGroup>
-
   <ItemGroup>
     <None Update="Resources\config\static-options.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
