+++ src-v2/NTS.TagRfidControllers/Hardware/ILogger.cs
@@ -1,0 +1,6 @@
+﻿namespace Rfid.Hardware;
+
+public interface ILogger
+{
+   void Log(string name, string error);
+}
