+++ src-v2/NTS.TagRfidControllers/NTS.TagRfidControllers.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Reference Include="Vup">
+      <HintPath>..\Judge\NTS.Judge\Vup.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+
+</Project>
